Fix 3 QA minors: mutex poison recovery, monotonic ETA, u64 return type
M1: Bounce output_path mutex now uses unwrap_or_else(into_inner)
    instead of silently ignoring poison — all 3 lock sites fixed
M2: Bounce ETA/speed uses monotonic Instant-based clock (BOUNCE_EPOCH)
    instead of SystemTime — immune to NTP/DST clock jumps
M3: engine_track_get_compensation_delay returns u64 (was u32)
    eliminates theoretical truncation on extreme sessions

+++ flutter_ui/lib/src/rust/native_ffi.dart
@@ -3632,7 +3632,7 @@ class NativeFFI {
 
     // Delay compensation
     _trackReportLatency = _lib.lookupFunction<Void Function(Uint64, Uint32), void Function(int, int)>('engine_track_report_latency');
-    _trackGetCompensationDelay = _lib.lookupFunction<Uint32 Function(Uint64), int Function(int)>('engine_track_get_compensation_delay');
+    _trackGetCompensationDelay = _lib.lookupFunction<Uint64 Function(Uint64), int Function(int)>('engine_track_get_compensation_delay');
     _setDelayCompensationEnabled = _lib.lookupFunction<Void Function(Int32), void Function(int)>('engine_set_delay_compensation_enabled');
     _isDelayCompensationEnabled = _lib.lookupFunction<Int32 Function(), int Function()>('engine_is_delay_compensation_enabled');
